+++ pioProjects/biolab_test1/include/oled.h
@@ -1,0 +1,9 @@
+void sendBitmap(const uint8_t *bitmap, uint8_t w, uint8_t h);
+void sendString(String toSend);
+void bootingPrint(void);
+void rebootingPrint(void);
+void pleaseWaitPrint(void);
+void testdrawcircle(void);
+void testdrawrect(void);
+void testdrawline();
+void clearAll();
