--- pioProjects/biolab_test1/include/test.h
@@ -1,1 +1,0 @@
-void testing(int x);
