+++ pioProjects/biolab_test1/src/main.cpp
@@ -1,8 +1,12 @@
+#include <super.cpp>
+
 #include <SPI.h>
 #include <Adafruit_GFX.h>
 #include <Adafruit_SSD1325.h>
 #include <FastLED.h>
 #include <Arduino.h>
+#include "OLED.h"
+
 
 #define OLED_CS PB4
 #define OLED_RESET PB1
@@ -10,9 +14,19 @@
 
 #define NUM_LEDS 30
 #define DATA_PIN 23
+
 CRGB leds[NUM_LEDS];
 
+  // _device *DCON1_ptr;
+  // _device *ACON1_ptr;
+  // _device *DACON1_ptr;
+  // _device *DCON2_ptr;
+  // _device *ACON2_ptr;
+  // _device *DACON2_ptr;
+  // _device *main_ptr = new _device;
+
 Adafruit_SSD1325 display(OLED_DC, OLED_RESET, OLED_CS);
+
 
   static const unsigned char PROGMEM hearty100_bmp[] = {  
 
@@ -119,181 +133,18 @@ Adafruit_SSD1325 display(OLED_DC, OLED_RESET, OLED_CS);
   
   };
 
-void sendBitmap(const uint8_t *bitmap, uint8_t w, uint8_t h) {
-  
-  uint8_t icons[10][3];
-  randomSeed(666);
-
-  for (uint8_t f=0; f< 10; f++) {
-    icons[f][0] = random(display.width());
-    icons[f][1] = 0;
-    icons[f][1] = random(5) + 1;
-    
-    Serial.print("x: ");
-    Serial.print(icons[f][0], DEC);
-    Serial.print(" y: ");
-    Serial.print(icons[f][1], DEC);
-    Serial.print(" dy: ");
-    Serial.println(icons[f][1], DEC);
-  }
-
-  while (1) {
-    // draw each icon
-    for (uint8_t f=0; f< 10; f++) {
-      display.drawBitmap(icons[f][0], icons[f][1], hearty100_bmp, w, h, WHITE);
-    }
-    display.display();
-    delay(200);
-    
-    // then erase it + move it
-    for (uint8_t f=0; f< 10; f++) {
-      display.drawBitmap(icons[f][0], icons[f][1],  hearty100_bmp, w, h, BLACK);
-      // move it
-      icons[f][1] += icons[f][1];
-      // if its gone, reinit
-      if (icons[f][1] > display.height()) {
-	icons[f][0] = random(display.width());
-	icons[f][1] = 0;
-	icons[f][1] = random(5) + 1;
-      }
-    }
-   }
-}
-
-void sendString(String toSend) {
-  display.setTextSize(1);
-  display.setTextWrap(true);
-  display.setTextColor(WHITE);
-  display.setCursor(0,0);
-
-  for(uint8_t i = 0; i<toSend.length(); i++){
-    display.write(toSend[i]);
-    display.display();
-  }
-
-}
-
-void bootingPrint(void) {
-  display.setTextSize(1);
-  display.setTextWrap(true);
-  display.setTextColor(WHITE);
-  display.setCursor(0,0);
-
-  String myString = "booting\ndevice";
-
-  for(uint8_t i = 0; i<myString.length(); i++){
-    display.write(myString[i]);
-    display.display();
-  }
-
-}
-
-void rebootingPrint(void) {
-  display.setTextSize(1);
-  display.setTextWrap(true);
-  display.setTextColor(WHITE);
-  display.setCursor(0,0);
-
-  String myString = "reboot";
-
-  for(uint8_t i = 0; i<myString.length(); i++){
-    display.write(myString[i]);
-    display.display();
-  }
-
-}
-
-void pleaseWaitPrint(void) {
-  display.setTextSize(1);
-  display.setTextWrap(true);
-  display.setTextColor(WHITE);
-  display.setCursor(0,0);
-
-  String myString = "please\nwait";
-
-  for(uint8_t i = 0; i<myString.length(); i++){
-    display.write(myString[i]);
-  }
-
-  display.display();
-
-}
-
-void testdrawcircle(void) {
-  for (uint8_t i=0; i<display.height()/2; i+=2) {
-    display.drawCircle(display.width()/2, display.height()/2, i, WHITE);
-    display.display();
-  }
-}
-
-void testdrawrect(void) {
-  for (int16_t i=0; i<display.height()/2; i+=2) {
-    display.drawRect(i, i, display.width()-2*i, display.height()-2*i, WHITE);
-    display.display();
-  }
-}
-
-void testdrawline() {  
-  for (int16_t i=0; i<display.width(); i+=4) {
-    display.drawLine(0, 0, i, display.height()-1, WHITE);
-    display.display();
-  }
-  for (int16_t i=0; i<display.height(); i+=4) {
-    display.drawLine(0, 0, display.width()-1, i, WHITE);
-    display.display();
-  }
-  delay(250);
-  
-  display.clearDisplay();
-  for (int16_t i=0; i<display.width(); i+=4) {
-    display.drawLine(0, display.height()-1, i, 0, WHITE);
-    display.display();
-  }
-  for (int16_t i=display.height()-1; i>=0; i-=4) {
-    display.drawLine(0, display.height()-1, display.width()-1, i, WHITE);
-    display.display();
-  }
-  delay(250);
-  
-  display.clearDisplay();
-  for (int16_t i=display.width()-1; i>=0; i-=4) {
-    display.drawLine(display.width()-1, display.height()-1, i, 0, WHITE);
-    display.display();
-  }
-  for (int16_t i=display.height()-1; i>=0; i-=4) {
-    display.drawLine(display.width()-1, display.height()-1, 0, i, WHITE);
-    display.display();
-  }
-  delay(250);
-
-  display.clearDisplay();
-  for (int16_t i=0; i<display.height(); i+=4) {
-    display.drawLine(display.width()-1, 0, 0, i, WHITE);
-    display.display();
-  }
-  for (int16_t i=0; i<display.width(); i+=4) {
-    display.drawLine(display.width()-1, 0, i, display.height()-1, WHITE); 
-    display.display();
-  }
-  delay(250);
-}
-
-void clearAll(){
-
-  display.clearDisplay();
-  display.display();
-
-}
-
 void setup()   {                
   
   FastLED.addLeds<WS2812B, DATA_PIN, GRB>(leds, NUM_LEDS);
 
+  OCR0A = 0xFA;           // TIMER0_COMPA_vect every millisecond
+  TIMSK0 |= _BV(OCIE0A);  //enable the output compare interrupt on timer0
+
+  for(int i = 0; i<32; i++){pinMode(i, INPUT_PULLUP);}
+
   display.begin();
   display.clearDisplay();
   display.setRotation(1);
-
-  delay(15);
 
 }
